created separete file for create rentals
