+++ user_choice.rb
@@ -2,6 +2,7 @@ require './app'
 require './menu'
 require './create_person'
 require './create_book'
+require './create_rental'
 
 # rubocop:disable Metrics/CyclomaticComplexity
 def user_choice(choice)
